Saving progress. Deleting deprecated code.
