+++ data/test_build_spec_lib/scenario_a_missed_cleavages/output/minimal_mc1.poin/build_log.txt
@@ -3,11 +3,11 @@
 Spectral Library Building Process
 ==========================================================================================
 
-Starting library build at: 2025-08-28T12:26:55.284
+Starting library build at: 2025-08-29T12:40:30.752
 Output directory: /Users/nathanwamsley/Projects/Pioneer.jl/test/Routines/BuildSpecLib/../../../data/test_build_spec_lib/scenario_a_missed_cleavages/output/minimal_mc1.poin
 
 Detecting fragment bounds...
-Fragment m/z range: Pioneer.FragBoundModel(ImmutablePolynomial(150.0f0), ImmutablePolynomial(2000.0f0))
+Fragment m/z range: FragBoundModel(ImmutablePolynomial(150.0f0), ImmutablePolynomial(2000.0f0))
 Precursor m/z range: (0.0f0, 5000.0f0)
 
 Starting fragment prediction workflow...
