+++ data/test_build_spec_lib/scenario_c_multifile/output/multifile_test.poin/build_log.txt
@@ -3,73 +3,14 @@
 Spectral Library Building Process
 ==========================================================================================
 
-Starting library build at: 2025-08-28T12:20:50.733
+Starting library build at: 2025-08-29T12:40:34.100
 Output directory: /Users/nathanwamsley/Projects/Pioneer.jl/test/Routines/BuildSpecLib/../../../data/test_build_spec_lib/scenario_c_multifile/output/multifile_test.poin
 
 Detecting fragment bounds...
-Fragment m/z range: Pioneer.FragBoundModel(ImmutablePolynomial(150.0f0), ImmutablePolynomial(2020.0f0))
+Fragment m/z range: FragBoundModel(ImmutablePolynomial(150.0f0), ImmutablePolynomial(2020.0f0))
 Precursor m/z range: (390.0f0, 1010.0f0)
 
 Starting fragment prediction workflow...
 Using prediction model: altimeter
 
 Preparing chronologer workflow...
-Predicting retention times...
-
-Predicting fragment ion intensities...
-
-Building final library indices...
-
-==========================================================================================
-Library Building Performance Report
-==========================================================================================
-
-Detailed Step Analysis:
-------------------------------------------------------------------------------------------
-# Precursors # Fragments  # Targets    # Decoys     
-------------------------------------------------------------------------------------------
-       52.00      3482.00        26.00        26.00 
-
-Detailed Step Analysis:
-------------------------------------------------------------------------------------------
-Step                                     Time (s)     Memory (GB)  GC Time (s)  GC %        
-------------------------------------------------------------------------------------------
-Collision Energy Setup                   0.00         0.00         0.00         0.0         
-File Verification                        0.00         0.00         0.00         0.0         
-Directory Setup                          0.00         0.00         0.00         0.0         
-Model Validation                         0.00         0.00         0.00         0.0         
-Parameter Loading                        0.00         0.00         0.00         0.0         
-Fragment Bound Detection                 0.00         0.00         0.00         0.0         
-Chronologer Preparation                  0.53         0.01         0.00         0.0         
-Chronologer Output Processing            0.57         0.01         0.05         8.9         
-Retention Time Prediction                0.78         0.01         0.00         0.0         
-Index Building                           0.78         0.02         0.05         5.8         
-Fragment Prediction                      1.00         0.01         0.00         0.0         
-Prediction Processing                    1.05         0.02         0.05         4.4         
-------------------------------------------------------------------------------------------
-TOTAL                                    4.72         0.08         0.14         3.0         
-
-Performance Summary:
-------------------------------------------------------------------------------------------
-Total Runtime: 0.08 minutes
-Total Garbage Collection Time: 0.14 seconds
-Number of Steps: 12
-Average Step Runtime: 0.39 seconds
-
-Memory Usage Summary:
-------------------------------------------------------------------------------------------
-Total Memory Allocated: 0.08 GB
-Peak  Memory Usage: 1.64 GB
-Total Available System Memory: 48.0 GB
-
-Process Information:
-------------------------------------------------------------------------------------------
-Number of Threads: 10
-Julia Version: 1.11.3
-System: arm64-apple-darwin24.0.0
-
-==========================================================================================
-
-Library building completed at: 2025-08-28T12:20:55.459
-
-Library location: /Users/nathanwamsley/Projects/Pioneer.jl/test/Routines/BuildSpecLib/../../../data/test_build_spec_lib/scenario_c_multifile/output/multifile_test.poin
